Some small modifications.
UnVals to UnDVs to be consistent with published papers.

modified the function names list so that in later versions of pin (2.9 onwards) .plt and other functions should not appear at the output.

modified the dot file printing such that the values are printed in separte lines on the edges. this is useful in complex applications and allows dot to produce more readable output.

removed the _ starting restriction from the unkown function name filter implementation, especially because KLT feature tracking library has functions (of our interest) name starting with underscores.